// src/app/shared/types/api-types/bidm-work-order-dto.interface.ts
import { BidmDocumentDTOId } from './bidm-document-dto-id.interface';
import { BidmDocumentDTO } from './bidm-document-dto.interface';

export interface BidmWorkOrderDTO {
  projectId?: string;
  woId?: string;
  bidProjectId?: string;
  bidWoId?: string;
  bsdProjectId?: string;
  invId?: string;
  woCode?: string;
  woSource?: string;
  woSourceText?: string;
  woType?: string;
  woDescription?: string;
  woAssignedTo?: string;
  woAdditionalWork?: string;
  woComments?: string;
  woLinkType?: string;
  taskPerformedBy?: string;
  bireItemFamilyCode?: string;
  bireItemVariantCode?: string;
  bireItemChapter?: string;
  bireItemSection?: string;
  bireItemSubject?: string;
  bireItemSheet?: string;
  bireItemMarks?: string;
  bsdeProjectNumber?: number;
  assetPn?: string;
  assetSn?: string;
  woScheduledStartDate?: Date;
  woScheduledEndDate?: Date;
  underWarranty?: boolean;
  warrantyProvider?: string;
  woPerformedBy?: string;
  woPerformedOn?: Date;
  woPerformedWithEsign?: boolean;
  woCheckedBy?: string;
  woCheckedOn?: Date;
  woCheckedWithEsign?: boolean;
  purchaseRequestCode?: string;
  salesRequestCode?: string;
  bidtWorkCenterId?: number;
  bidtWarehouseId?: number;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
  bidmDocumentAddedList?: BidmDocumentDTO[];
  bidmDocumentDeletedList?: BidmDocumentDTOId[];
  bidmDocumentUpdatedList?: BidmDocumentDTO[];
  calculatedStatus?: string;
  bidmProjectNumber?: number;
}