// src/app/shared/types/api-types/bidt-procurement-request-dto.interface.ts
// //////////////////////////////////////////////////////////////////////////
// DO NOT EDIT THIS FILE
//
// It is automatically generated by generate-dtos.js
// //////////////////////////////////////////////////////////////////////////

import { BidmWorkOrderDTO } from './bidm-work-order-dto.interface';
import { BidoCustomerDTO } from './bido-customer-dto.interface';
import { BidoEquipmentDTO } from './bido-equipment-dto.interface';
import { BidtDeliveryFolderDTO } from './bidt-delivery-folder-dto.interface';
import { BidtDocumentDTO } from './bidt-document-dto.interface';
import { BidtProcReqItemDTO } from './bidt-proc-req-item-dto.interface';
import { BidtProcurementRequestIdDTO } from './bidt-procurement-request-id-dto.interface';
import { BidtSiteDTO } from './bidt-site-dto.interface';
import { BidtWarehouseDTO } from './bidt-warehouse-dto.interface';

export interface BidtProcurementRequestDTO extends BidtProcurementRequestIdDTO {
  bidtDeliveryFolders?: BidtDeliveryFolderDTO[];
  bidtDocuments?: BidtDocumentDTO[];
  bidtProcReqItems?: BidtProcReqItemDTO[];
  bidtSiteByBidtSiteIssueId?: number;
  bidtSiteByBidtSiteReceiptId?: number;
  bidtWarehouseByBidtWarehouseIssueId?: number;
  bidtWarehouseByBidtWarehouseReceiptId?: number;
  bidmWorkOrderBidMandant?: string;
  bidmWorkOrderProjectId?: string;
  bidmWorkOrderMandant?: string;
  bidmWorkOrderWoId?: string;
  calendarPotential?: Date;
  contract?: string; // 50
  criticality?: string; // 50
  criticalityReason?: string; // 1000
  potentialUnit?: string; // 50
  procCode?: string; // 50
  procComment?: string; // 1000
  procStatus?: string; // 50
  procType?: string; // 50
  requestedDate?: Date;
  requestedPotential?: string;
  bidoCustomerCustomerCode?: string;
  bidoCustomerMandant?: string;

  bidoCustomer?: BidoCustomerDTO;
  bidmWorkOrder?: BidmWorkOrderDTO;
  bidtSiteIssue?: BidtSiteDTO;
  bidtSiteReceipt?: BidtSiteDTO;
  bidtWarehouseIssue?: BidtWarehouseDTO;
  bidtWarehouseReceipt?: BidtWarehouseDTO;

  statusDate?: Date;
  statusState?: string; // 50
  statusUser?: string; // 100
  bidoEquipmentDTO?: BidoEquipmentDTO;
}